Identifying and fixing tracking errors
This script is able to identify tracking errors. Generally, tracking errors can occur due to transition of new born bacteria and divisions of existing bacteria (assigning more than two bacteria to a mother after 1.5/3 min which is biologically impossible). The both errors lead to mislabeling (please note that in CP all cells of a lineage have the same label)